# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserSignUpResponse": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"preferences": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"userId",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserSignUpResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserSignUpResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FailResponse_null_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAIL"
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"data": {
								"type": "number",
								"enum": [
									null
								],
								"nullable": true
							},
							"reason": {
								"type": "string"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"reason",
							"errorCode"
						],
						"type": "object"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"birth": {
						"type": "string",
						"format": "date-time"
					},
					"address": {
						"type": "string"
					},
					"detailAddress": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"email",
					"password",
					"name",
					"gender",
					"birth",
					"phoneNumber",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"restaurantId": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"user": {
						"properties": {
							"name": {
								"type": "string"
							}
						},
						"required": [
							"name"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"restaurantId",
					"userId",
					"rating",
					"content",
					"createdAt",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ReviewItem"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ReviewListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/ReviewListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"missionId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"mission": {
						"properties": {
							"restaurant": {
								"properties": {
									"name": {
										"type": "string"
									}
								},
								"required": [
									"name"
								],
								"type": "object"
							},
							"dueDate": {
								"type": "string"
							},
							"mealPrice": {
								"type": "number",
								"format": "double"
							},
							"point": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"restaurant",
							"dueDate",
							"mealPrice",
							"point"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"userId",
					"missionId",
					"status",
					"createdAt",
					"updatedAt",
					"mission"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserMissionItem"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserMissionListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserMissionListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewAddResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"restaurantId": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"restaurantId",
					"userId",
					"rating",
					"content",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ReviewAddResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/ReviewAddResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewAddRequest": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"rating",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ReviewAddRequest_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/ReviewAddRequest"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"restaurantId": {
						"type": "number",
						"format": "double"
					},
					"point": {
						"type": "number",
						"format": "double"
					},
					"mealPrice": {
						"type": "number",
						"format": "double"
					},
					"dueDate": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"restaurant": {
						"properties": {
							"name": {
								"type": "string"
							}
						},
						"required": [
							"name"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"restaurantId",
					"point",
					"mealPrice",
					"dueDate",
					"createdAt",
					"updatedAt",
					"restaurant"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/MissionItem"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_MissionListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/MissionListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionStartResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"missionId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"userId",
					"missionId",
					"status",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_MissionStartResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/MissionStartResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FailResponse": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAIL"
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"data": {
								"type": "number",
								"enum": [
									null
								],
								"nullable": true
							},
							"reason": {
								"type": "string"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"reason",
							"errorCode"
						],
						"type": "object"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionStartRequest": {
				"properties": {
					"missionId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"missionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionAddResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"restaurantId": {
						"type": "number",
						"format": "double"
					},
					"point": {
						"type": "number",
						"format": "double"
					},
					"mealPrice": {
						"type": "number",
						"format": "double"
					},
					"dueDate": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"restaurantId",
					"point",
					"mealPrice",
					"dueDate",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_MissionAddResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/MissionAddResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionAddRequest": {
				"properties": {
					"point": {
						"type": "number",
						"format": "double"
					},
					"mealPrice": {
						"type": "number",
						"format": "double"
					},
					"dueDate": {
						"type": "string"
					}
				},
				"required": [
					"point",
					"mealPrice",
					"dueDate"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"paths": {
		"/users/signup": {
			"post": {
				"operationId": "HandleUserSignUp",
				"responses": {
					"200": {
						"description": "회원가입 결과",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserSignUpResponse_"
								}
							}
						}
					},
					"409": {
						"description": "중복된 이메일 에러",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailResponse_null_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "U001",
												"reason": "이미 사용 중인 이메일입니다.",
												"data": null
											},
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "회원가입 API",
				"summary": "회원가입을 처리하는 엔드포인트입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignUpRequest"
							}
						}
					}
				}
			}
		},
		"/users/guest": {
			"get": {
				"operationId": "HandleGuestPage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "게스트 페이지",
				"summary": "로그인 없이 접근 가능한 게스트 페이지입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/login": {
			"get": {
				"operationId": "HandleLoginPage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "로그인 페이지",
				"summary": "인증이 필요한 페이지 시 리디렉션되는 로그인 페이지입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/mypage": {
			"get": {
				"operationId": "HandleMypage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "마이페이지",
				"summary": "로그인한 유저만 접근 가능한 마이페이지입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/set-login": {
			"get": {
				"operationId": "HandleSetLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/set-logout": {
			"get": {
				"operationId": "HandleSetLogout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/{userId}/reviews": {
			"get": {
				"operationId": "HandleListUserReview",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewListResponse_"
								}
							}
						}
					}
				},
				"description": "유저 리뷰 목록 조회 API",
				"summary": "특정 유저가 작성한 리뷰 목록을 반환하는 엔드포인트입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/missions": {
			"get": {
				"operationId": "HandleListUserMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionListResponse_"
								}
							}
						}
					}
				},
				"description": "유저 미션 목록 조회 API",
				"summary": "특정 유저가 수행중인 미션 목록을 반환하는 엔드포인트입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "HandleStartMission",
				"responses": {
					"200": {
						"description": "미션 시작 결과",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_MissionStartResponse_"
								}
							}
						}
					},
					"409": {
						"description": "이미 진행 중이거나 완료한 미션 — AlreadyChallengingMissionError (M002)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "M002",
												"reason": "이미 해당 미션을 진행 중이거나 완료했습니다.",
												"data": null
											},
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"summary": "미션 시작을 요청하는 엔드포인트입니다.",
				"tags": [
					"UserMission"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MissionStartRequest"
							}
						}
					}
				}
			}
		},
		"/restaurants/{restaurantId}/reviews": {
			"post": {
				"operationId": "HandleAddReview",
				"responses": {
					"200": {
						"description": "리뷰 등록 결과",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewAddResponse_"
								}
							}
						}
					}
				},
				"summary": "리뷰 등록을 처리하는 엔드포인트입니다.",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "restaurantId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewAddRequest"
							}
						}
					}
				}
			}
		},
		"/restaurants/{resId}/reviews": {
			"get": {
				"operationId": "HandleListRestaurantReviews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewListResponse_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 식당 — RestaurantNotFoundError (R001)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailResponse_null_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "R001",
												"reason": "해당 식당을 찾을 수 없습니다.",
												"data": null
											},
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"summary": "식당별 리뷰를 조회하는 엔드포인트입니다.",
				"tags": [
					"Restaurants"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "resId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/restaurants/{resId}/missions": {
			"get": {
				"operationId": "HandleListRestaurantMissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_MissionListResponse_"
								}
							}
						}
					}
				},
				"summary": "식당별 미션을 조회할 수 있는 엔드포인트입니다.",
				"tags": [
					"Restaurants"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "resId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/restaurants/{restaurantId}/missions": {
			"post": {
				"operationId": "HandleAddMission",
				"responses": {
					"200": {
						"description": "미션 추가 결과",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_MissionAddResponse_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 유저 또는 미션 — NotFoundError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailResponse"
								}
							}
						}
					}
				},
				"summary": "식당에 미션을 추가할 수 있는 엔드포인트입니다.",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "restaurantId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MissionAddRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}